Rework how we handle metrics and introduce event tracking
The attempt to implement event metrics tracking revealed that I needed to rethink how to track metrics in general.

We need to store the tracked metrics per address on the phone, while we only want to send generalized metrics per area. Once the local storage of a device is cleared we can't aggregate on earlier metrics but start a new "metrics" session.
For that reason I needed to introduce a TrackingSessionStore persisting local tracking data in local storage, starting new sessions if needed.

This commit might contain some drive-by fixes where the application bugged out.

refs CAC_FRONTEND-567 CAC_FRONTEND-547