Solved issue of weird performance bug
The mapping of the the vee-validate field is not compatible to the events emitted by the primevue dropdown.

We probably should search a more elegant way to fix this, but mapping the onInput listeners to the @input event of the dropdown, while only listening for the value solves the problem.