Use clustering on map to handle large datasets CAC_FRONTEND-23
Use the new clustering endpoint to display clusters on higher zoomlevels. While zooming in the clusters are recalculated until eventually there are less than 100 in the bounding box to display.
If on a zoom level higher than 15 there are still more than 100 events display a warning to better use the list view