For short-lived session omit saving the refresh token CAC_SERVICE-16
After some consideration I deemed this mechanism sufficient for now. If you deselect the checkbox "Stay logged in" the refresh token will not be persisted, limiting the session
to the exiry time of the access token (10 hours for now). After that a stolen secret becomes stale.

A minor nuisance can occur if you actually use the token until it expires you will be logged out while workin, with the need of reentering your credentials, but i think the effort to
implement a mechanism to have expiring refresh tokens is not worth it just now.