Revert backbutton behavior to resemble native 'back'
Using forward routing in a back button can lead to all kinds of trouble messing up the browser history, so that the behavior of the
soft-back button is inconsistent with the hard back button. Until we figure out a universal concept we should stick with resembling the native behavior which is go(-1).