Change toolchain to quasar framework
This commit changes the existing codebase to be embedded in the quasar toolchain including the quasar framework.
Adjustments were to match the eslint / tsconfig of the quasar framework as well as catering to the proposed directory structure of the toolchain.

Key benifits:

* Quasar UI Framework is _way_ more cross-platform friendly as Ionic only targets ios / android, quasar components work on desktop well, too, which we ultimately aim for with this app
* Quasar toolchain includes helpers and guides to compile out PWA and capacitor-wrapped applications
* Features like SSR and tree pre-configured shaking are compelling arguments

Plan would be to fade out ionic as quickly as possible to ultimately ditch it.