CAC_FRONTEND-122 Remove zooming on marker drag and page change
The zoom was always execute even when the marker was moved/dragged in
the visible area, also the zoom was executed again on going forward to
the area editing page. These obsolete zooms have been removed. The zoom
is now just executed on search of a location or on first load of the
map.