CAC_FRONTEND-297 Remove direct binding of `onMoveActivePoster`
If `onMoveActivePoster` is bind directly to `touchmove`,
then no matter if poster is selected or not the app tries to move a
poster, which leads to awkward navigation.

To make a valid movement of a poster, a user needs to execute an
`touchstart` on a poster (identified as `activePosterLayerId`) before a
movement of poster should triggered, this means `onMoveActivePoster` and
respectively `onUpActivePoster` should be bind after an `touchstart`,
which is already like this. In the function `onTouchStartActivePoster`
the two functions are bind

```
map.value.on('touchmove', onMoveActivePoster)
map.value.once('touchend', onUpActivePoster)
```

So it's just necessary to remove the binding in the very first place.