CAC_FRONTEND-330 Fix TS2571 Object is of type 'unknown'
With updating of typescript to version 4.4 the catch variable is defaulting
to Type `unknown` instead of `any`.
(see [Announcing-typescript-4-4](https://devblogs.microsoft.com/typescript/announcing-typescript-4-4/#use-unknown-catch-variables))
Which was leading to error:
```
TS2571: Object is of type 'unknown'.
```
To mitigate this, we need we narrow all errors to an `AxiosError` or basic
Error`, depends on what we expect in the catch.

For the `AxiosError`s we're using [`axios.isAxiosError`](https://github.com/axios/axios#typescript),
wrapped inside `apiClient.isApiClientError`which checks if catch variable is an
axios error and if yes narrows it to Type `AxiosError`.

relevant-links:
- https://devblogs.microsoft.com/typescript/announcing-typescript-4-4/#use-unknown-catch-variables
- https://github.com/axios/axios#typescript
- https://www.typescriptlang.org/docs/handbook/2/narrowing.html#using-type-predicates