CAC_FRONTEND-335 Fix distinctBy function
The `distinctBy` function was not matching it's goal.

E.g. we pass an array of objects and a function
```js
array = [{id:1, name: 'a'}, {id:1, name: 'a'}, {id:2, name: 'b'}]
fn = (item) => item.id
```

After first round of execution the for-loop, `distinct` would look like:
```
distinct = [{id:1, name: 'a'}]
```

in the second round it would check:
```
!distinct.includes(fn(item))
```
which is
```
![{id:1, name: 'a'}].includes(1)
```
which will always be false and also every subsequent check.
So nothing is filtered. To fix this also the items of distinct array needs to
be mapped to property returned by `fn`