Remove unused css class
Due to an update of quasar something related to Sass behavior have been update, which was throwing the deprecation warning:

```
28  │ ┌   Button {
29  │ │     padding: 3px 6px;
30  │ │   }
    │ └─── nested rule
... │
32  │     margin-top: 6px;
    │     ^^^^^^^^^^^^^^^ declaration
    ╵
    src/map/popup/EventPopupContents.vue 32:3  root stylesheet

Deprecation Warning: Sass's behavior for declarations that appear after nested
rules will be changing to match the behavior specified by CSS in an upcoming
version. To keep the existing behavior, move the declaration above the nested
rule. To opt into the new behavior, wrap the declaration in `& {}`.

More info: https://sass-lang.com/d/mixed-decls
```

To fix this it was possible to remove the css class, since it wasn't used anymore