Fix declaration warning addressing nested rules
Origin of the fix was the following deprecation warning:

```
15  │ ┌     h3 {
16  │ │       display: inline-block;
17  │ │       margin: 0 1rem 0 0;
18  │ │     }
    │ └─── nested rule
... │
21  │       display: flex;
    │       ^^^^^^^^^^^^^ declaration
    ╵
    src/components/CollapsibleFilters.vue 21:5  root stylesheet

Deprecation Warning: Sass's behavior for declarations that appear after nested
rules will be changing to match the behavior specified by CSS in an upcoming
version. To keep the existing behavior, move the declaration above the nested
rule. To opt into the new behavior, wrap the declaration in `& {}`.

More info: https://sass-lang.com/d/mixed-decls
```
To fix this it was required to move declaration above the nested rule.