Fix setting for margin
It seems that due to an update of quasar/sass the way the selector `* {` was evaluated changed. Instead of just applying to just the 'first generation' of childrens it was now applied to each children and grandchildren, which changed the way the elements/buttons have looked liked. To fix this the selector is set ot `> *` to just select
the first level/generation of children.