#70 Shrink `SELECT_AREA_SET` page
If no collections are present the dialog showing the options
to search areas by name and to pick an area from old events
doens't need to take 100% of height. In this case it's fine if
the size of the dialog is smaller and just encompass the size of
these two options.