#79 fic sending of completion note
Before this fix:
When user/devices have been offline while putting
in the first metrics, the completion note wasn't send
to the server successfully. Even when later on more metrics
have been collected the completion note wasn't send to server.
Since it was just done on the intial tracking of metrics.

To fix this we're checking if the completion note is existing or
if the property of completion note `completed` is set to `false`.
Then we're know that the completion note wasn't send yet.
see `hasBecomeCompleted`

Something similar was able to happen when an address
has become Incomple see `hasBecomeIncompleted`.