#94 Limit number size to 99
The html attribute "maxlength" is not working for inputs with
`type="number"`. Due to this switch to default type of input and added
inputmode="numeric" to ensure that at least on mobile just the numberic
keyboard is shown. We also added some error toast for desktops if user
puts in wrong data with string.