#103 Migrate custom user store to pinia store
The old store was always initialized before the App was loaded.
We want the store to be just loaded and initialized when it is used,
Like this it's possible to execute a migration script in the boot phase
of the app. Before I tried to add a migration script to the boot scope
of quasar, but the store was always executed/initialized in beforehand,
but the migration of local storage needs to be run before store is
initialized.