#114 Refresh event area when updating assignees
When people are getting assigned to an event,
the meta data of event area is also getting updated.
Explicitly the `has_assignee` is getting updated.
Since eventarea information is not getting the update automatically, we
need to refetch the data, so that also the FE with and the style of the
layer are getting updated accordingl.